Refactor and supply required comments
