+++ Saule/Http/BulkExtRouteAttributeAttribute.cs
@@ -1,18 +1,27 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
-using System.Net.Http;
 using System.Web.Http.Routing;
 
 namespace Saule.Http
 {
+    /// <summary>
+    /// Custom route attribute to support negotiating the same route depending on content type to support the bulk extension.
+    /// </summary>
     public class BulkExtRouteAttributeAttribute : RouteFactoryAttribute
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BulkExtRouteAttributeAttribute"/> class.
+        /// </summary>
+        /// <param name="template">Route name</param>
+        /// <param name="multiple">Sets whether this route is standard JSON API or bulk extension enabled</param>
         public BulkExtRouteAttributeAttribute(string template, bool multiple)
             : base(template)
         {
             Multiple = multiple;
         }
 
+        /// <summary>
+        /// Gets overriden constraints handling to select route based on the attribute.
+        /// </summary>
         public override IDictionary<string, object> Constraints
         {
             get
@@ -31,41 +40,9 @@ namespace Saule.Http
             }
         }
 
+        /// <summary>
+        /// Gets a value indicating whether this route is standard JSON API or bulk extension enabled
+        /// </summary>
         public bool Multiple { get; private set; }
-    }
-
-    internal class ContentTypeConstraint : IHttpRouteConstraint
-    {
-        public ContentTypeConstraint(string allowedMediaType)
-        {
-            AllowedMediaType = allowedMediaType;
-        }
-
-        public string AllowedMediaType { get; private set; }
-
-        public bool Match(HttpRequestMessage request, IHttpRoute route, string parameterName, IDictionary<string, object> values, HttpRouteDirection routeDirection)
-        {
-            if (routeDirection == HttpRouteDirection.UriResolution)
-            {
-                return GetMediaHeader(request) == AllowedMediaType;
-            }
-            else
-            {
-                return true;
-            }
-        }
-
-        private string GetMediaHeader(HttpRequestMessage request)
-        {
-            IEnumerable<string> headerValues;
-            if (request.Content.Headers.TryGetValues("Content-Type", out headerValues) && headerValues.Count() == 1)
-            {
-                return headerValues.First();
-            }
-            else
-            {
-                return "application/vnd.api+json";
-            }
-        }
     }
 }
